Move worker-error-reporting.js to web-worker/.
